// src/app/main/model/contact/contact.ts
export class Contact {
    _id: string;
    contactID: string;
    creator: string;
    dateCreate: string;
    dateModify: string;
    contactName: string;
    englishName: string;
    address: string;
    birthday: string;
    telephone: string;
    identifyCard: string;
    extNo: string;
    joiningDate: string;
    position: string;
    group: string;
    department: string;
    knowledge: string;
    email: string;
    emailPassword: string;
    photo: any;
    photoSize: string;
    approveAmount: number;
    exceptionApproveBy: string;
    saleTargetUSD: number;
    saleTargetLC: number;
    percentBonus: number;
    percentFixedProfit: number;
    userName: string;
    password: string;
    accountUsernameReference: string;
    disable: boolean;
    stopWorking: boolean;
    stopWorkingDate: string;
    isLocalUserInOffice: boolean;
    hideSearch: boolean;
    accessRight: string;
    accessDescription: string;

    constructor() {
        this.contactID = null;
        this.creator = null;
        this.dateCreate = null;
        this.dateModify = null;
        this.contactName = null;
        this.englishName = null;
        this.address = null;
        this.birthday = null;
        this.telephone = null;
        this.identifyCard = null;
        this.extNo = null;
        this.joiningDate = null;
        this.position = null;
        this.group = null;
        this.department = null;
        this.knowledge = null;
        this.email = null;
        this.emailPassword = null;
        this.photo = null;
        this.photoSize = null;
        this.approveAmount = null;
        this.exceptionApproveBy = null;
        this.saleTargetUSD = null;
        this.saleTargetLC = null;
        this.percentBonus = null;
        this.percentFixedProfit = null;
        this.userName = null;
        this.password = null;
        this.accountUsernameReference = null;
        this.disable = false;
        this.stopWorking = false;
        this.stopWorkingDate = null;
        this.isLocalUserInOffice = false;
        this.hideSearch = false;
        this.accessRight = null;
        this.accessDescription = null;
    }
}